refactor(ui): Extract function to create plugin options schema
The functions to create the form schema for the create and edit forms
for plugin templates were almost identical, so extract the common code
to a helper function.
