+++ ui/src/routes/admin/plugins/$pluginType/$pluginId/edit-template/$templateName/index.tsx
@@ -28,7 +28,7 @@ import {
 import { Loader2 } from 'lucide-react';
 import { ChangeEvent } from 'react';
 import { Resolver, useForm } from 'react-hook-form';
-import { z, ZodType } from 'zod';
+import { z } from 'zod';
 
 import { PluginOption, PluginOptionTemplate, PluginOptionType } from '@/api';
 import {
@@ -63,24 +63,11 @@ import { ApiError } from '@/lib/api-error';
 import { queryClient } from '@/lib/query-client';
 import { toast, toastError } from '@/lib/toast';
 import { getPluginTypeLabel } from '@/lib/types';
-import { pluginOptionTypeToZodType } from '../../-helpers.ts';
+import { buildPluginOptionsFormShape } from '../../-helpers.ts';
 import { Route as LayoutRoute } from '../../../../route.tsx';
 
 function buildFormSchema(options: Array<PluginOption>) {
-  const shape: Record<string, ZodType> = {};
-  for (const opt of options) {
-    let schema = pluginOptionTypeToZodType(opt.type);
-    if (opt.isNullable) {
-      schema = schema.nullable();
-    }
-    if (!opt.isRequired) {
-      schema = schema.optional();
-    }
-    shape[opt.name] = schema;
-    shape[`${opt.name}_isFinal`] = z.boolean().default(false);
-    shape[`${opt.name}_isNotSet`] = z.boolean().default(false);
-  }
-  return z.object(shape);
+  return z.object(buildPluginOptionsFormShape(options));
 }
 
 function parseStoredValue(
